server tracking pt. 2, preparation for storing that data into DB

+++ src/main/java/com/offliner/rust/rust_offliner/interfaces/IServerDao.java
@@ -15,5 +15,5 @@ public interface IServerDao extends CrudRepository<ServerEntity, Integer> {
     long countAllByCurrentlyTrackedIsTrue();
 
     @Query(nativeQuery = true, value = "SELECT server_id FROM `server` WHERE tracked = 1")
-    List<Long> getAllByServerIdAndCurrentlyTrackedIsTrue();
+    List<Integer> getAllByServerIdAndCurrentlyTrackedIsTrue();
 }
